added some toasts and made some changes to make adding possible
adding still does not work

+++ app/src/main/java/com/e/truehomemobile/models/apartment/Apartment.kt
@@ -7,7 +7,7 @@ class Apartment(
     var apartmentStreet: String,
     var apartmentStreetNumber: String,
     var apartmentZipCode: String,
-    var apartmentPrice: Int,
+    var apartmentPrice: Double,
     var apartmentDescription: String,
     var apartmentImage: String
 ) {
